+++ .github/workflows/ci.yml
@@ -11,7 +11,10 @@ env:
 
 jobs:
   fmt:
-    runs-on: "ubuntu-latest"
+    strategy:
+      matrix:
+        os: ["ubuntu-latest", "windows-latest"]
+    runs-on: ${{ matrix.os }}
     steps:
       - name: Free Disk Space (Ubuntu)
         uses: insightsengineering/disk-space-reclaimer@v1
@@ -27,18 +30,27 @@ jobs:
       - uses: actions-rust-lang/setup-rust-toolchain@v1
         with:
           components: rustfmt, clippy
+          rustflags: '' #Disable.  By default this action sets environment variable is set to -D warnings.  We manage this in the Makefile
       - name: Setup build env
-        run: ./scripts/setup.sh
+        run: |
+          os=$(echo "$RUNNER_OS" | tr '[:upper:]' '[:lower:]')
+          ./scripts/setup-$os.sh
         shell: bash
       - run: 
           # needed to run rustfmt in nightly toolchain
           rustup toolchain install nightly --component rustfmt
+      - name: Set environment variables for Windows
+        if: runner.os == 'Windows'
+        run: |
+          # required until standalong is implemented for windows (https://github.com/WasmEdge/wasmedge-rust-sdk/issues/54)
+          echo "WASMEDGE_LIB_DIR=C:\Program Files\WasmEdge\lib" >> $env:GITHUB_ENV
+          echo "WASMEDGE_INCLUDE_DIR=C:\Program Files\WasmEdge\include" >> $env:GITHUB_ENV
       - name: Run checks
         run: make check
   build:
     strategy:
       matrix:
-        os: ["ubuntu-20.04", "ubuntu-22.04"]
+        os: ["ubuntu-20.04", "ubuntu-22.04", "windows-latest"]
     runs-on: ${{ matrix.os }}
     steps:
       - name: Free Disk Space (Ubuntu)
@@ -53,18 +65,25 @@ jobs:
           docker-images: true
       - name: "check cgroup version"
         run: "mount | grep cgroup"
+        if: runner.os == 'Linux'
       - uses: actions/checkout@v3
       - uses: actions-rust-lang/setup-rust-toolchain@v1
         env:
           RUST_CACHE_KEY_OS: ${{ matrix.os }}
-      - name: Setup build env
-        run: ./scripts/setup.sh
-        shell: bash
-      - name: Build
-        uses: actions-rs/cargo@v1
         with:
-          command: build
-          args: --all --verbose
+           rustflags: '' #Disable.  By default this action sets environment variable is set to -D warnings.  We manage this in the Makefile
+      - name: Setup build env
+        run: |
+          os=$(echo "$RUNNER_OS" | tr '[:upper:]' '[:lower:]')
+          ./scripts/setup-$os.sh
+        shell: bash
+      - name: Set environment variables for Windows
+        if: runner.os == 'Windows'
+        run: |
+          echo "WASMEDGE_LIB_DIR=C:\Program Files\WasmEdge\lib" >> $env:GITHUB_ENV
+          echo "WASMEDGE_INCLUDE_DIR=C:\Program Files\WasmEdge\include" >> $env:GITHUB_ENV
+      - name: Build
+        run: make build
       - name: Validate docs
         run: ./scripts/validate-docs.sh
       - name: Run tests
@@ -95,7 +114,7 @@ jobs:
       - name: setup rust-wasm target
         run: rustup target add wasm32-wasi
       - name: Setup build env
-        run: ./scripts/setup.sh
+        run: ./scripts/setup-linux.sh
         shell: bash
       - name: run
         run: make test/k8s
@@ -132,7 +151,7 @@ jobs:
       - name: setup rust-wasm target
         run: rustup target add wasm32-wasi
       - name: Setup build env
-        run: ./scripts/setup.sh
+        run: ./scripts/setup-linux.sh
         shell: bash
       - name: run
         run: make test/k3s
