+++ crates/containerd-shim-wasm/src/sandbox/testutil.rs
@@ -74,11 +74,20 @@ macro_rules! function {
     }};
 }
 
+#[cfg(unix)]
 use caps::{CapSet, Capability};
 pub use function;
 
 /// Determines if the current process has the CAP_SYS_ADMIN capability in its effective set.
 pub fn has_cap_sys_admin() -> bool {
-    let caps = caps::read(None, CapSet::Effective).unwrap();
-    caps.contains(&Capability::CAP_SYS_ADMIN)
+    #[cfg(unix)]
+    {
+        let caps = caps::read(None, CapSet::Effective).unwrap();
+        caps.contains(&Capability::CAP_SYS_ADMIN)
+    }
+
+    #[cfg(windows)]
+    {
+        false
+    }
 }
