+++ crates/containerd-shim-wasmedge/Cargo.toml
@@ -17,10 +17,12 @@ oci-spec = { workspace = true, features = ["runtime"] }
 thiserror = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-nix = { workspace = true }
 libc = { workspace = true }
+
+[target.'cfg(unix)'.dependencies]
 libcontainer = { workspace = true }
 dbus = { version = "*", optional = true }
+nix = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3.7"
