+++ crates/containerd-shim-wasmedge/src/lib.rs
@@ -1,8 +1,12 @@
 pub mod error;
-pub mod executor;
+
 pub mod instance;
 pub mod oci_utils;
 
+#[cfg(unix)]
+pub mod executor;
+
+#[cfg(unix)]
 #[cfg(test)]
 mod test {
     use std::os::unix::prelude::OsStrExt;
