+++ crates/containerd-shim-wasmtime/Cargo.toml
@@ -32,11 +32,13 @@ cap-std = { workspace = true }
 oci-spec = { workspace = true, features = ["runtime"] }
 thiserror = { workspace = true }
 serde_json = { workspace = true }
+serde = { workspace = true }
+libc = { workspace = true }
+
+[target.'cfg(unix)'.dependencies]
 nix = { workspace = true }
 libcontainer = { workspace = true }
 dbus = { version = "*", optional = true }
-serde = { workspace = true }
-libc = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3.7"
