+++ crates/containerd-shim-wasmtime/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod error;
-pub mod executor;
 pub mod instance;
 pub mod oci_wasmtime;
+
+#[cfg(unix)]
+pub mod executor;
