+++ rustfmt.toml
@@ -1,4 +1,4 @@
-newline_style = "Unix"
+newline_style = "Native"
 unstable_features = true # Cargo fmt now needs to be called with `cargo +nightly fmt`
 group_imports = "StdExternalCrate" # create three groups for std, external and local crates
 # Merge imports from the same module
